requirements.txt
python_dotenv==1.0.1
streamlit==1.37.0
google-generativeai
PyPDF2
fpdf==1.7.2
langchain
markdown2
reportlab
beautifulsoup4
langchain-google-genai